+++ gpu/src/easy_app.cpp
@@ -97,7 +97,7 @@ int main(int argc, char **argv) {
     num_run++;
   }
 
-  double total_time = 0;
+  float total_time = 0;
 
   std::cout << std::endl << "====Summary====" << std::endl;
   std::cout << "Raw data copy avg time for 1 frame : " << AvgVec(beamformer.memcpy_time) << " ms." << std::endl;
